Export nuget.exe path
This will explicitly set which nuget.exe will be used which is critical to how packaging works on mac.  Some versions of nuget.exe will produce corrupt .zip files.  This should cause the cake build scripts usage of nuget.exe to be explicitly the one downloaded by this bootstrapper script