Added a m2repository dependency diff util
Each time we update the sdk version we bind against, we have to re-evaluate the dependency chain between the native libraries so that we can mimick them in our bindings (for nuget packages and the binding projects themselves).

Traditionally we’ve used an internal utility to help list the dependency graph from a maven repository.

This workbook file will help determine changes in the dependency graph between versions of the native libraries.

The idea is to use this workbook and manually inspect and implement the dependency graph changes.  The workbook makes it much easier and quicker to figure out what those changes are.