Added type forwarder AppIndexing package
We previously removed Xamarin.GooglePlayServices.AppIndexing as google no longer ships a play-services-appindexing, but instead they ship firebase-appindexing with all the old types (and more) that were previously in the old package.

This caused breaks in apps where they referenced libraries depending on Xamarin.GooglePlayServices.AppIndexing from an older version wanting to upgrade their play services packages to a new version.

This new package will ship an assembly with only type forwarded statements in it to allow old libraries to continue to work and see the types in their new location.