Add extra nuspec templates to processing
The Aar listing is currently used to loop through items to process nuspec template files for, however with the typeforwarded lib, we don’t have an aar to extract, so there’s no aar info listing for it and it wouldn’t be processed.  This adds the processing to other arbitrary files.