Added target to include org.apache.http.legacy.jar
In proguard invocations so we don’t get warnings which result in failures since classes in this jar can’t be found, even though they aren’t needed (that or someone would have to add `-dontwarn org.apache.**` to their proguard rules, but this is friendlier.

Also, added `_GenerateAndroidResourceDir` to the `BeforeTargets` for the target that writes out the res filewrites, as it wasn’t getting run in design time builds.