[msbuild] fix for $(AndroidUseAapt2)
If used in a Xamarin.Android project with aapt2,
`@(GoogleServicesJson)` items cause targets to run on every build:

    Building target "_ConvertLibraryResourcesCases" completely.
    Input file "obj\Debug.stamp" does not exist.

So one problem, is Xamarin.Android has some logic to decide the path
of the `.stamp` file on directories that are normally in
`obj\Debug\lp\1\`:

https://github.com/xamarin/xamarin-android/blob/b6f7f0e40ac9e500acd59ba61f045f7dde5dcc99/src/Xamarin.Android.Build.Tasks/Tasks/CollectNonEmptyDirectories.cs#L30

We need to make fixes in Xamarin.Android, but a workaround is to set
the `%(StampFile)` metadata on `@(LibraryResourceDirectories)`. This
will make our targets rerun aapt2 only when a `@(GoogleServicesJson)`
file changes.