[utilities.cake] Only look for namespaces in "obj" assemblies. (#766)
Context: https://github.com/xamarin/AndroidX/pull/722

For some reason, `0.2.0-alpha4` of `Xamarin.Legacy.Sdk` causes package dependencies to be placed in the `/bin` directory whereas previously they were not.  Our published namespace script looks for all `.dll`'s in the `/generated` directory, which previously did not include dependencies.  This resulted in adding namespaces that were in dependencies and not just the assemblies packaged in this repository.

To fix this, we can restrict ourselves to looking for assemblies in the `/obj` directory, which does not contain dependency assemblies.