Add extension to Asset Delivery (#842)
It adds the `AssetPackStateUpdateListenerWrapper` to the `Xamarini.Google.Android.Play.Asset.Delivery` nuget package to help users deal with ondemand asset packs. 
This class is also n the Xamarin.Google.Android.Play.Core package, but that package has been depricated  and appears to have issues on API 34.

Remove the old `AssetPackStateUpdateListenerWrapper` from `Xamarini.Google.Android.Play.Core` as it will conflict with the one that will be in `Xamarini.Google.Android.Play.Asset.Delivery`. Do the same for the `SplitInstallStateUpdateListenerWrapper` this will now be in `Xamarini.Google.Android.Play.Feature.Delivery`.

Add a sample app called `BuildAllPlayDotNet` which will use the Asset Delivery API and the wrapper class to make sure it is compiled into the nuget package.